Continued implementation of SuperByteMatrix

+++ src/net/maizegenetics/util/SuperByteMatrixBuilder.java
@@ -13,7 +13,7 @@ public class SuperByteMatrixBuilder {
     }
 
     public static SuperByteMatrix getInstance(int rows, int columns) {
-        if (rows * columns > Integer.MAX_VALUE) {
+        if ((long) rows * (long) columns > Integer.MAX_VALUE) {
             return new SuperByteMatrixMultiple(rows, columns);
         } else {
             return new SuperByteMatrixSingle(rows, columns);
